Enable student to view presentation points
- Added presentation points for graded presentations to students' course statistics
- restricted presentationsWeight to [0, 99], since 100 would lead to an unbelievable amount of reachable presentation points 🤯

+++ src/main/java/de/tum/in/www1/artemis/service/exam/ExamService.java
@@ -292,7 +292,7 @@ public class ExamService {
             Double achievedPointsOfSource = 0.0;
             PlagiarismVerdict verdict = null;
             Integer presentationScoreThreshold = null;
-            Integer achievedPresentationScore = null;
+            Double achievedPresentationScore = null;
             if (result != null) {
                 achievedPointsOfSource = result.achievedPoints();
                 verdict = result.mostSeverePlagiarismVerdict();
